updated the two trace visualization scripts
it's recommended to use trace2google.py, which analyze a trace
dump and output the trace event in Google's Trace Event Format

because the trace_vis frontend uses timeline-js, which is not
intended for visualizing tracing result. so it can be very slow
and unresponsive when working with large trace.

speedscope (https://github.com/jlfwong/speedscope) or Chrome's
about://tracing can be used for viewing the Trace Event Format

Signed-off-by: Jiayang Li <admin@0x10c.pw>